submodules/moragents_dockers/agents/requirements.txt
llama-cpp-python==0.2.90
sentencepiece==0.2.0
protobuf==5.27.2
scikit-learn==1.5.1
flask==2.2.2
Werkzeug==2.2.2
flask-cors==4.0.1
web3==6.20.1
pymupdf==1.22.5
faiss-cpu==1.8.0.post1
langchain-text-splitters==0.2.2
langchain-core==0.2.24
langchain-community==0.2.10
langchain_experimental==0.3.2
tweepy==4.14.0
